Changed prices to doubles instead of ints

+++ challenge/Basket.java
@@ -38,8 +38,8 @@ public class Basket {
     return customerBasket.size();
   }
 
-  public int getTotalCost() {
-    int totalCost = 0;
+  public double getTotalCost() {
+    double totalCost = 0.0;
     for (Product item : customerBasket)
       totalCost = totalCost + item.getCost(); {
         return totalCost;
